requirements.txt
bs4                             ==0.0.1
openai                          ==0.27.2
ebooklib                        ==0.18
rich                            ==13.3.2
opencc-python-reimplemented     ==0.1.7
